# Optimizations: Node memory layout and Triangle struct