Bringing some optimizations to earcut:

* Change the memory layout of the Node object to reduce memory size (~30% smaller)
* Improve runtime speed by caching triangle coords in a struct that fits within a cache line


after (ns) | before (ns) | change | name
--- | --- | --- | ---
3639.09 | 3502.92 | 3.89% | `bad_hole`
309.23 | 299.4 | 3.28% | `building`
141.93 | 138.94 | 2.15% | `degenerate`
4982.33 | 5105.03 | -2.40% | `dude`
191.56 | 193.82 | -1.17% | `empty_square`
397645.67 | 440010.5 | -9.63% | `water`
314034.67 | 340645.83 | -7.81% | `water2`
13054.3 | 13607.89 | -4.07% | `water3`
1232.5 | 1251.44 | -1.51% | `water3b`
80921.71 | 77718.75 | 4.12% | `water4`
6449166.5 | 6848604.5 | -5.83% | `water_huge`
14539417 | 15063062.5 | -3.48% | `water_huge2`